:wrench: Refactor request validation handling by moving it to the storage server and removing it from vector routes
